Move set UA from _sdk/utils to _utils/user_agent_utils
