+++ src/promptflow/promptflow/_cli/_pf/entry.py
@@ -23,15 +23,12 @@ from promptflow._cli._pf._connection import add_connection_parser, dispatch_conn
 from promptflow._cli._pf._flow import add_flow_parser, dispatch_flow_commands  # noqa: E402
 from promptflow._cli._pf._run import add_run_parser, dispatch_run_commands  # noqa: E402
 from promptflow._cli._pf._tool import add_tool_parser, dispatch_tool_commands  # noqa: E402
-from promptflow._cli._pf.help import show_privacy_statement, show_welcome_message  # noqa: E402
 from promptflow._cli._pf._upgrade import add_upgrade_parser, upgrade_version  # noqa: E402
+from promptflow._cli._pf.help import show_privacy_statement, show_welcome_message  # noqa: E402
 from promptflow._cli._user_agent import USER_AGENT  # noqa: E402
-from promptflow._sdk._utils import (  # noqa: E402
-    get_promptflow_sdk_version,
-    print_pf_version,
-    setup_user_agent_to_operation_context,
-)
+from promptflow._sdk._utils import get_promptflow_sdk_version, print_pf_version  # noqa: E402
 from promptflow._utils.logger_utils import get_cli_sdk_logger  # noqa: E402
+from promptflow._utils.user_agent_utils import setup_user_agent_to_operation_context  # noqa: E402
 
 # get logger for CLI
 logger = get_cli_sdk_logger()
