+++ src/promptflow/promptflow/_sdk/_pf_client.py
@@ -8,12 +8,12 @@ from typing import Any, Dict, List, Union
 
 from .._constants import USER_AGENT_OVERRIDE_KEY
 from .._utils.logger_utils import get_cli_sdk_logger
+from .._utils.user_agent_utils import ClientUserAgentUtil, setup_user_agent_to_operation_context
 from ..exceptions import ErrorTarget, UserErrorException
 from ._configuration import Configuration
 from ._constants import MAX_SHOW_DETAILS_RESULTS, ConnectionProvider
 from ._load_functions import load_flow
 from ._user_agent import USER_AGENT
-from ._utils import ClientUserAgentUtil, setup_user_agent_to_operation_context
 from .entities import Run
 from .entities._eager_flow import FlexFlow
 from .operations import RunOperations
