+++ src/promptflow/tests/sdk_cli_azure_test/e2etests/test_azure_cli_perf.py
@@ -5,11 +5,11 @@ from typing import Callable
 from unittest import mock
 
 import pytest
+from sdk_cli_azure_test.recording_utilities import is_replay
 
 from promptflow._cli._user_agent import USER_AGENT as CLI_USER_AGENT  # noqa: E402
 from promptflow._sdk._telemetry import log_activity
-from promptflow._sdk._utils import ClientUserAgentUtil
-from sdk_cli_azure_test.recording_utilities import is_replay
+from promptflow._utils.user_agent_utils import ClientUserAgentUtil
 
 FLOWS_DIR = "./tests/test_configs/flows"
 DATAS_DIR = "./tests/test_configs/datas"
